# backend/src/main/kotlin/dev/martinschwarz/jokes/domain/JokeCategory.kt
package dev.martinschwarz.jokes.domain

/** Coarse grouping used to filter the catalogue and to draw a themed random joke. */
enum class JokeCategory {
    GENERAL,
    PUN,
    DAD,
    PROGRAMMING,
}