# backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Joke API
  description: A small catalogue of jokes — browse, draw a random one, contribute your own.
  version: 0.0.1
servers:
  - url: http://localhost:8080
paths:
  /jokes:
    get:
      summary: List jokes, newest first
      tags:
        - Jokes
      operationId: listJokes
      parameters:
        - name: category
          in: query
          required: false
          description: Restrict the page to a single category; omit for all categories.
          schema:
            $ref: '#/components/schemas/JokeCategory'
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of jokes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JokePage'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Add a joke to the catalogue
      tags:
        - Jokes
      operationId: createJoke
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJokeRequest'
      responses:
        '201':
          description: The created joke; `Location` carries its canonical URL.
          headers:
            Location:
              description: URL of the created joke.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Joke'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: An identical joke already exists in this category.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /jokes/random:
    get:
      summary: Draw a random joke
      tags:
        - Jokes
      operationId: getRandomJoke
      parameters:
        - name: category
          in: query
          required: false
          description: Draw from a single category; omit to draw from all.
          schema:
            $ref: '#/components/schemas/JokeCategory'
      responses:
        '200':
          description: A randomly selected joke
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Joke'
        '404':
          description: The catalogue (or the requested category) holds no jokes.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /jokes/{id}:
    get:
      summary: Fetch a single joke
      tags:
        - Jokes
      operationId: getJoke
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The joke
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Joke'
        '404':
          description: No joke with this id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  responses:
    BadRequest:
      description: The request violated a validation rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    JokeCategory:
      type: string
      description: Coarse grouping used for filtering and for drawing a themed random joke.
      enum:
        - GENERAL
        - PUN
        - DAD
        - PROGRAMMING
    Joke:
      type: object
      required:
        - id
        - content
        - category
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        category:
          $ref: '#/components/schemas/JokeCategory'
        createdAt:
          type: string
          format: date-time
    JokePage:
      type: object
      description: A slice of the catalogue plus the counters a client needs to paginate.
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Joke'
        page:
          type: integer
          format: int32
          description: Zero-based index of the returned page.
        size:
          type: integer
          format: int32
          description: Maximum number of items per page.
        totalElements:
          type: integer
          format: int64
          description: Total jokes matching the filter across all pages.
        totalPages:
          type: integer
          format: int32
    CreateJokeRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 500
        category:
          $ref: '#/components/schemas/JokeCategory'
          default: GENERAL
    Problem:
      type: object
      description: RFC 9457 problem detail, as produced by Spring's `ProblemDetail` support.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri